refac: move run in transaction to cli utils
