+++ migra/src/clients/mod.rs
@@ -35,29 +35,3 @@ pub use self::mysql::Client as MysqlClient;
 pub mod sqlite;
 #[cfg(feature = "sqlite")]
 pub use self::sqlite::Client as SqliteClient;
-
-pub fn run_in_transaction<TrxFnMut>(client: &mut AnyClient, trx_fn: TrxFnMut) -> MigraResult<()>
-where
-    TrxFnMut: FnOnce(&mut AnyClient) -> MigraResult<()>,
-{
-    client
-        .begin_transaction()
-        .and_then(|_| trx_fn(client))
-        .and_then(|res| client.commit_transaction().and(Ok(res)))
-        .or_else(|err| client.rollback_transaction().and(Err(err)))
-}
-
-pub fn should_run_in_transaction<TrxFnMut>(
-    client: &mut AnyClient,
-    is_needed: bool,
-    trx_fn: TrxFnMut,
-) -> MigraResult<()>
-where
-    TrxFnMut: FnOnce(&mut AnyClient) -> MigraResult<()>,
-{
-    if is_needed {
-        run_in_transaction(client, trx_fn)
-    } else {
-        trx_fn(client)
-    }
-}
