+++ migra/src/lib.rs
@@ -9,6 +9,5 @@ pub mod fs;
 pub mod managers;
 pub mod migration;
 
-pub use clients::{run_in_transaction, should_run_in_transaction};
 pub use errors::{Error, MigraResult as Result, StdResult};
 pub use migration::Migration;
