+++ migra_cli/src/commands/apply.rs
@@ -1,7 +1,6 @@
 use crate::app::App;
 use crate::database;
 use crate::opts::ApplyCommandOpt;
-use migra::should_run_in_transaction;
 
 pub(crate) fn apply_sql(app: &App, cmd_opts: &ApplyCommandOpt) -> migra::StdResult<()> {
     let config = app.config()?;
@@ -21,14 +20,14 @@ pub(crate) fn apply_sql(app: &App, cmd_opts: &ApplyCommandOpt) -> migra::StdResu
         .map(std::fs::read_to_string)
         .collect::<Result<Vec<_>, _>>()?;
 
-    should_run_in_transaction(
+    database::should_run_in_transaction(
         &mut client,
         cmd_opts.transaction_opts.single_transaction,
         |client| {
             file_contents
                 .iter()
                 .try_for_each(|content| {
-                    should_run_in_transaction(
+                    database::should_run_in_transaction(
                         client,
                         !cmd_opts.transaction_opts.single_transaction,
                         |client| client.apply_sql(content),
