+++ migra_cli/src/commands/downgrade.rs
@@ -1,7 +1,6 @@
 use crate::app::App;
 use crate::database;
 use crate::opts::DowngradeCommandOpt;
-use migra::should_run_in_transaction;
 use std::cmp;
 
 pub(crate) fn rollback_applied_migrations(
@@ -33,7 +32,7 @@ pub(crate) fn rollback_applied_migrations(
         })
         .collect::<Result<Vec<_>, _>>()?;
 
-    should_run_in_transaction(
+    database::should_run_in_transaction(
         &mut client,
         opts.transaction_opts.single_transaction,
         |client| {
@@ -42,7 +41,7 @@ pub(crate) fn rollback_applied_migrations(
                 .try_for_each(|(migration_name, content)| {
                     if all_migrations.contains_name(migration_name) {
                         println!("downgrade {}...", migration_name);
-                        should_run_in_transaction(
+                        database::should_run_in_transaction(
                             client,
                             !opts.transaction_opts.single_transaction,
                             |client| client.run_downgrade_migration(migration_name, &content),
