+++ migra_cli/src/database.rs
@@ -42,3 +42,29 @@ pub fn create_client_from_config(config: &Config) -> migra::StdResult<AnyClient>
     )
     .map_err(From::from)
 }
+
+pub fn run_in_transaction<TrxFnMut>(client: &mut AnyClient, trx_fn: TrxFnMut) -> migra::Result<()>
+where
+    TrxFnMut: FnOnce(&mut AnyClient) -> migra::Result<()>,
+{
+    client
+        .begin_transaction()
+        .and_then(|_| trx_fn(client))
+        .and_then(|res| client.commit_transaction().and(Ok(res)))
+        .or_else(|err| client.rollback_transaction().and(Err(err)))
+}
+
+pub fn should_run_in_transaction<TrxFnMut>(
+    client: &mut AnyClient,
+    is_needed: bool,
+    trx_fn: TrxFnMut,
+) -> migra::Result<()>
+where
+    TrxFnMut: FnOnce(&mut AnyClient) -> migra::Result<()>,
+{
+    if is_needed {
+        run_in_transaction(client, trx_fn)
+    } else {
+        trx_fn(client)
+    }
+}
